pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "traitprint"
version = "0.2.0"
description = "Local-first career identity vault for the agent era. MCP server for AI agents."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "DataViking Technologies", email = "hello@dataviking.tech" },
]
keywords = ["mcp", "career", "identity", "vault", "local-first", "ai-agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "mcp",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
]
import = [
    "pypdf>=4.0",
    "python-docx>=1.0",
]

[project.scripts]
traitprint = "traitprint.cli:cli"

[project.urls]
Homepage = "https://traitprint.com"
Repository = "https://github.com/DataViking-Tech/traitprint"
Issues = "https://github.com/DataViking-Tech/traitprint/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/traitprint"]

[tool.ruff]
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
